+++ CHANGELOG.md
@@ -16,7 +16,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [2.5.4] - 2019-01-09
+
+### Fixed
+
 - Normalization of strings that failed to decompose under certain conditions (#106)
+- Throw only on errors, not on errorcode that has `WARNING` in name if `throwOnWarnings == false`
 
 ## [2.5.3] - 2018-12-17